perf: reduce first load js size of landing page by 75kB
