+++ app/web-app/src/components/proposal-form.tsx
@@ -29,7 +29,7 @@ import { erc20ABI, fundingVaultABI } from '@/blockchain/constants';
 import { parseUnits } from 'viem';
 import { useWeb3FormSubmit } from '@/hooks/use-web3-form-submit';
 import { Web3SubmitButton } from '@/components/web3-submit-button';
-import { isValidURL } from '@/lib/utils';
+import { isValidURL } from '@/lib/is-valid-url';
 
 interface ProposalFormProps {
     fundingVault: FundingVault;
