+++ app/web-app/src/lib/get-token-name.ts
@@ -1,0 +1,17 @@
+import { config as wagmiConfig } from '@/wagmi/config';
+import { erc20ABI } from '@/blockchain/constants';
+import { readContract } from '@wagmi/core';
+
+export async function getTokenName(address: string): Promise<string> {
+    try {
+        const result = await readContract(wagmiConfig, {
+            address: address as `0x${string}`,
+            abi: erc20ABI,
+            functionName: 'symbol',
+        });
+        return result as string;
+    } catch (err) {
+        console.log('[GET_TOKEN_NAME_ERROR]', err, address);
+        return 'Unknown';
+    }
+}
