+++ app/web-app/src/lib/is-valid-url.ts
@@ -1,0 +1,10 @@
+export function isValidURL(url: string | undefined): boolean {
+    try {
+        if (url) {
+            new URL(url);
+        }
+        return true;
+    } catch (error) {
+        return false;
+    }
+}
