+++ app/web-app/src/types/index.d.ts
@@ -8,3 +8,9 @@ declare interface SmartContract {
     address: string;
     abi: any;
 }
+
+declare module 'lucide-react/dist/esm/icons/*' {
+    import { LucideIcon } from 'lucide-react';
+    const icon: LucideIcon;
+    export default icon;
+}
