+++ app/web-app/tsconfig.json
@@ -21,6 +21,6 @@
             "@/*": ["./src/*"]
         }
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts","index.d.ts"],
     "exclude": ["node_modules"]
 }
